requirements.txt
streamlit
langchain-google-genai
langchain-community
pypdf
beautifulsoup4
python-dotenv